More updates. ver 1.4.5
Testing found that I needed separate variable day/night twilightzone
diff variables.  Also added debut log feature that makes it a lot easier
to tune settings.  Hard to duplicate twilight artificially with correct
timing so it takes a while to get everything right.  Will add a section
on tuning.  I have managed to automate a lot of the variables.  Not the
best of code but it does the job.  It took 3 complete algorithm changes
but the 4th time looks good.  Hope it has a successful winter run.